feat: extract ProjectInfo component
